done with first practice method

+++ 02 - Java Tutorial/SpeedConverter/src/Main.java
@@ -1,7 +1,7 @@
 public class Main {
     public static void main(String[] args) {
 
-        double miles = SpeedConverter.toMilesPerHour(10.5);
+        long miles = SpeedConverter.toMilesPerHour(10.5);
         System.out.println(miles);
         SpeedConverter.printConversion(10.5);
     }
